+++ changes.rst
@@ -4,6 +4,26 @@ Nake changes log
 
 Changes for `nake <https://github.com/fowlmouth/nake>`_.
 
+v1.6, 2015-04-26
+----------------
+
+* `Silences compilation warnings <https://github.com/fowlmouth/nake/pull/22>`_.
+  This change requires Nim 0.10.2 or better.
+* `Fixed broken documentation links, added website with generated docs
+  <https://github.com/fowlmouth/nake/issues/27>`_.
+* `Added test suite <https://github.com/fowlmouth/nake/pull/26>`_.
+* `Moved nake procs and macros to nakelib module which can be imported without
+  turning the importer into a nakefile
+  <https://github.com/fowlmouth/nake/issues/19>`_.
+* `Removed prefix letter from types
+  <https://github.com/fowlmouth/nake/issues/29>`_.
+* `Reduced nake verbosity, added silent shell procs
+  <https://github.com/fowlmouth/nake/issues/20>`_.
+* `Adds nake's timestamp to nakefile rebuild dependency list
+  <https://github.com/fowlmouth/nake/issues/23>`_.
+* `DATWPL license was changed to MIT
+  <https://github.com/fowlmouth/nake/pull/37>`_.
+
 v1.4, 2014-12-30
 ----------------